requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Env loading
python-dotenv==1.0.1

# Frontend
streamlit==1.35.0
requests==2.32.3

# Computer vision
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# Face detection
# installed separately in Dockerfile before this file to avoid resolver issues
insightface==0.7.3
onnxruntime==1.18.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0